refactor: extract event list component
