+++ src/app/page.tsx
@@ -2,6 +2,7 @@
 
 import EventForm from '@/components/EventForm';
 import DownloadButton from '@/components/DownloadButton';
+import EventList from '@/components/EventList';
 import { useState } from 'react';
 import { CalendarEvent } from '@/lib/icsGenerator';
 import { useTranslations } from 'next-intl';
@@ -20,24 +21,27 @@ export default function HomePage() {
 
   return (
     <main className='min-h-screen bg-gradient-to-br from-background to-primary/10 text-foreground px-4 py-10 transition-colors flex items-center justify-center'>
-      <div className='w-full max-w-2xl space-y-8 text-center animate-fade-in'>
-        <h1 className='text-3xl md:text-5xl font-bold'>{`📅 ${t('title')}`}</h1>
-        <p className='text-sm md:text-base opacity-80'>
-          Create a calendar file (.ics) for your events. Supports English,
-          日本語 & 한국어.
-        </p>
+      <div className='w-full max-w-4xl space-y-8 animate-fade-in'>
+        <div className='text-center space-y-2'>
+          <h1 className='text-3xl md:text-5xl font-bold'>{`📅 ${t('title')}`}</h1>
+          <p className='text-sm md:text-base opacity-80'>
+            Create a calendar file (.ics) for your events. Supports English,
+            日本語 & 한국어.
+          </p>
+        </div>
 
-        <EventForm
-          onAdd={handleAdd}
-          onRemove={handleRemove}
-          events={events}
-        />
+        <div className='flex flex-col md:flex-row gap-8 md:items-start'>
+          <EventForm onAdd={handleAdd} />
+          <EventList events={events} onRemove={handleRemove} />
+        </div>
+
         <DownloadButton events={events} />
 
-        <footer className='pt-6 text-xs opacity-60'>
+        <footer className='pt-6 text-xs opacity-60 text-center'>
           Made by Yunsu Bae - 다국어 지원 테스트 중
         </footer>
       </div>
     </main>
   );
 }
+
